+++ lib/oxc/lint/native.ex
@@ -22,7 +22,7 @@ defmodule OXC.Lint.Native do
     ),
     version: version
 
-  @spec lint(String.t(), String.t(), [String.t()], [{String.t(), String.t()}], boolean()) ::
+  @spec lint(iodata(), String.t(), [String.t()], [{String.t(), String.t()}], boolean()) ::
           {:ok, [map()]} | {:error, [String.t()]}
   def lint(_source, _filename, _plugins, _rules, _fix), do: :erlang.nif_error(:nif_not_loaded)
 end